Upload new csv
I compiled this dataset that contains information on all-time medals by country, number of participations in Games, the breakdown in Gold, Silver, Bronze, some averages and also some data on the countries like population, elevation, HDI (human development index) and a value for medals per million people. From this dataset we can extract some information on how and if the country of birth affects the chances for success.